Task3/Sleep Posture/requirements.txt
tensorflow
streamlit
mediapipe
opencv-python
numpy
Pillow
matplotlib
seaborn
scikit-learn